requirements.txt
Django==1.6.5
pinax-theme-bootstrap==5.4.0
django-user-accounts==1.0c9
metron==1.3.1
eventlog==0.8.0
django-jsonfield==0.9.12
python-social-auth==0.1.21
south==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
python_crontab==1.8.1
beautifulsoup4==4.3.2
django_tastypie==0.12.1
requests==2.3.0
MySQL-python==1.2.5
sqlalchemy==0.9.8
six==1.7.3
Pillow==2.5.3
cloudinary==1.0.18

